tools/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
pypdf>=3.17.1
unstructured>=0.10.30
faiss-cpu>=1.7.4
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
python-magic>=0.4.27
tqdm>=4.66.1
aiohttp>=3.9.1 